backend/requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
Flask-Migrate==2.2.1
pytest==3.7.1
factory-boy==2.11.1
inflection==0.3.1
marshmallow==3.0.0b13
flask-marshmallow==0.9.0
marshmallow-sqlalchemy==0.14.0
pytest-flask-sqlalchemy==1.0.0
psycopg2==2.7.5
flake8==3.5.0
coverage==4.5.1